+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add validation of duplicate responses to LookupPage admin
 - Fix 128: Use inline index subtitle in table of contents and pagination
+- Use PostgreSQL in Vagrant box and CI
 
 ## 37.3 (2021-06-11)