+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Add feature flags to disable page feedback and job alert forms
 - Add validation of duplicate responses to LookupPage admin
 - Fix 128: Use inline index subtitle in table of contents and pagination
 - Use PostgreSQL in Vagrant box and CI